a couple of checks for operations, of which I just realized that they where unsafe

+++ app/src/main/java/org/y20k/escapepod/helpers/FileHelper.kt
@@ -125,17 +125,20 @@ object FileHelper {
     /* Clears given folder - keeps given number of files */
     fun clearFolder(folder: File?, keep: Int, deleteFolder: Boolean = false) {
         if (folder != null && folder.exists()) {
-            val files = folder.listFiles()
-            val fileCount: Int = files.size
-            files.sortBy { it.lastModified() }
-            for (fileNumber in files.indices) {
-                if (fileNumber < fileCount - keep) {
-                    files[fileNumber].delete()
+            val files: Array<File>? = folder.listFiles()
+            if (files != null) {
+                val fileCount: Int = files.size
+                files.sortBy { it.lastModified() }
+                for (fileNumber in files.indices) {
+                    if (fileNumber < fileCount - keep) {
+                        files[fileNumber].delete()
+                    }
+                }
+                if (deleteFolder && keep == 0) {
+                    folder.delete()
                 }
             }
-            if (deleteFolder && keep == 0) {
-                folder.delete()
-            }
+
         }
     }
